Split out signing
