+++ lib/node.js
@@ -38,6 +38,8 @@ class Node extends EventEmitter {
     this._onData = onData.bind(this)
     this._onResult = onResult.bind(this)
     this._onFinished = onFinished.bind(this)
+    this._onSignAnnounce = onSignAnnounce.bind(this)
+    this._onSignUnannounce = onSignUnannounce.bind(this)
     this._onNoiseRequest = onNoiseRequest.bind(this)
 
     this._socket
@@ -52,6 +54,8 @@ class Node extends EventEmitter {
       .on('data', this._onData)
       .on('result', this._onResult)
       .on('finished', this._onFinished)
+      .on('signAnnounce', this._onSignAnnounce)
+      .on('signUnannounce', this._onSignUnannounce)
       .on('noiseRequest', this._onNoiseRequest)
       .alive()
       .handshake({
@@ -268,6 +272,18 @@ function onFinished (message) {
   const query = this._queries.get(message.id)
 
   if (query) query._onFinished(message)
+}
+
+function onSignAnnounce (message) {
+  const server = this._servers.get(message.publicKey)
+
+  if (server && server._signing) server._signing._onSignAnnounce(message)
+}
+
+function onSignUnannounce (message) {
+  const server = this._servers.get(message.publicKey)
+
+  if (server && server._signing) server._signing._onSignUnannounce(message)
 }
 
 function onNoiseRequest (message) {
