+++ lib/server.js
@@ -1,16 +1,10 @@
 const EventEmitter = require('events')
-const buffer = require('b4a')
-const sodium = require('sodium-universal')
 const SecretStream = require('@hyperswarm/secret-stream')
-const { encode } = require('compact-encoding')
-const { NS } = require('@hyperswarm/dht/lib/constants')
 
 const { Protocol } = require('./protocol')
+const { Signing } = require('./signing')
 const { Socket } = require('./socket')
 const { SocketSet } = require('./socket-set')
-
-const messages = require('./messages')
-const crypto = require('./crypto')
 
 class Server extends EventEmitter {
   constructor (node, socket, protocol, options = {}) {
@@ -24,6 +18,7 @@ class Server extends EventEmitter {
     this._custodial = options.custodial !== false
 
     this._connections = new SocketSet()
+    this._signing = null
     this._keyPair = null
     this._address = null
     this._listening = null
@@ -35,8 +30,6 @@ class Server extends EventEmitter {
     this._onConnection = onConnection.bind(this)
     this._onDestroy = onDestroy.bind(this)
     this._onData = onData.bind(this)
-    this._onSignAnnounce = onSignAnnounce.bind(this)
-    this._onSignUnannounce = onSignUnannounce.bind(this)
 
     this._socket
       .on('close', this._onClose)
@@ -44,8 +37,6 @@ class Server extends EventEmitter {
     this._protocol
       .on('destroy', this._onDestroy)
       .on('data', this._onData)
-      .on('signAnnounce', this._onSignAnnounce)
-      .on('signUnannounce', this._onSignUnannounce)
 
     this.ready = async function ready () {
       await this._protocol.ready()
@@ -60,6 +51,7 @@ class Server extends EventEmitter {
     if (this._listening) return this._listening
 
     this._keyPair = keyPair
+    this._signing = new Signing(this._protocol, keyPair)
 
     this._node._servers.add(keyPair.publicKey, this)
 
@@ -106,8 +98,6 @@ function onClose () {
   this._protocol
     .off('destroy', this._onDestroy)
     .off('data', this._onData)
-    .off('signAnnounce', this._onSignAnnounce)
-    .off('signUnannounce', this._onSignUnannounce)
 
   this.emit('close')
 }
@@ -161,59 +151,6 @@ function onData (message) {
   }
 }
 
-async function onSignAnnounce (message) {
-  const target = crypto.hash(this.publicKey)
-
-  const peer = {
-    publicKey: this.publicKey,
-    relayAddresses: message.relayAddresses
-  }
-
-  const data = signable(target, message.token, message.peerId, peer, NS.ANNOUNCE)
-
-  const signature = buffer.allocUnsafe(64)
-
-  sodium.crypto_sign_detached(signature, data, this._keyPair.secretKey)
-
-  await this._protocol.signature({
-    id: message.id,
-    signature
-  })
-}
-
-async function onSignUnannounce (message) {
-  const target = crypto.hash(this.publicKey)
-
-  const peer = {
-    publicKey: this.publicKey,
-    relayAddresses: message.relayAddresses
-  }
-
-  const data = signable(target, message.token, message.peerId, peer, NS.UNANNOUNCE)
-
-  const signature = buffer.allocUnsafe(64)
-
-  sodium.crypto_sign_detached(signature, data, this._keyPair.secretKey)
-
-  await this._protocol.signature({
-    id: message.id,
-    signature
-  })
-}
-
 function allowAll () {
   return false
-}
-
-function signable (target, token, id, peer, ns) {
-  const hash = buffer.allocUnsafe(32)
-
-  sodium.crypto_generichash_batch(hash, [
-    target,
-    id,
-    token,
-    encode(messages.peer, peer)
-  ], ns)
-
-  return hash
 }
